version 5.0 Enchance print functionalities, added purchase and issue list with item details - improve top header and users tab

+++ app/Http/Controllers/IssueController.php
@@ -86,7 +86,7 @@ class IssueController extends Controller
             // Stock availability check (audit-safe)
             foreach ($validated['lines'] as $i => $line) {
                 $itemId = (int) $line['item_id'];
-                $qty = (float) $line['quantity'];
+                $qty = (int) $line['quantity'];
 
                 $available = $stock->getAvailableStock($itemId);
