+++ database/migrations/2026_01_18_080001_create_issue_return_lines_table.php
@@ -13,9 +13,10 @@ return new class extends Migration {
             $table->foreignId('issue_line_id')->constrained('issue_lines')->cascadeOnDelete();
             $table->foreignId('item_id')->constrained('items');
             $table->string('specification')->nullable();
-            $table->decimal('issue_price', 15, 2)->default(0);
-            $table->decimal('quantity', 15, 3);
-            $table->decimal('line_total', 15, 2)->default(0);
+            // Business rule: integer quantities and prices
+            $table->unsignedInteger('issue_price')->default(0);
+            $table->unsignedInteger('quantity');
+            $table->unsignedBigInteger('line_total')->default(0);
             $table->timestamps();
         });
     }
